Refactor to own converter class
